refactor: Move cmd queries to the root level of cmd/aspect
