+++ cmd/aspect/aquery/aquery.go
@@ -7,10 +7,11 @@ Not licensed for re-use
 package aquery
 
 import (
+	"github.com/spf13/cobra"
+
 	"aspect.build/cli/pkg/aspect/query/aquery"
 	"aspect.build/cli/pkg/bazel"
 	"aspect.build/cli/pkg/ioutils"
-	"github.com/spf13/cobra"
 )
 
 func NewDefaultAQueryCmd() *cobra.Command {
