+++ cmd/aspect/cquery/BUILD.bazel
@@ -1,12 +1,12 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "aquery",
-    srcs = ["aquery.go"],
-    importpath = "aspect.build/cli/cmd/aspect/query/aquery",
+    name = "cquery",
+    srcs = ["cquery.go"],
+    importpath = "aspect.build/cli/cmd/aspect/cquery",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/aspect/query/aquery",
+        "//pkg/aspect/query/cquery",
         "//pkg/bazel",
         "//pkg/ioutils",
         "@com_github_spf13_cobra//:cobra",
