+++ cmd/aspect/query/query.go
@@ -7,10 +7,11 @@ Not licensed for re-use
 package query
 
 import (
+	"github.com/spf13/cobra"
+
 	"aspect.build/cli/pkg/aspect/query/query"
 	"aspect.build/cli/pkg/bazel"
 	"aspect.build/cli/pkg/ioutils"
-	"github.com/spf13/cobra"
 )
 
 func NewDefaultQueryCmd() *cobra.Command {
