+++ cmd/aspect/root/BUILD.bazel
@@ -9,13 +9,13 @@ go_library(
         "//cmd/docgen:__pkg__",
     ],
     deps = [
+        "//cmd/aspect/aquery",
         "//cmd/aspect/build",
         "//cmd/aspect/clean",
+        "//cmd/aspect/cquery",
         "//cmd/aspect/docs",
         "//cmd/aspect/info",
-        "//cmd/aspect/query/aquery",
-        "//cmd/aspect/query/cquery",
-        "//cmd/aspect/query/query",
+        "//cmd/aspect/query",
         "//cmd/aspect/run",
         "//cmd/aspect/test",
         "//cmd/aspect/version",
