+++ cmd/aspect/root/root.go
@@ -15,13 +15,13 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
+	"aspect.build/cli/cmd/aspect/aquery"
 	"aspect.build/cli/cmd/aspect/build"
 	"aspect.build/cli/cmd/aspect/clean"
+	"aspect.build/cli/cmd/aspect/cquery"
 	"aspect.build/cli/cmd/aspect/docs"
 	"aspect.build/cli/cmd/aspect/info"
-	"aspect.build/cli/cmd/aspect/query/aquery"
-	"aspect.build/cli/cmd/aspect/query/cquery"
-	"aspect.build/cli/cmd/aspect/query/query"
+	"aspect.build/cli/cmd/aspect/query"
 	"aspect.build/cli/cmd/aspect/run"
 	"aspect.build/cli/cmd/aspect/test"
 	"aspect.build/cli/cmd/aspect/version"
